Move has dimensions
